pyproject.toml
[project]
name = "kemopin"
version = "0.2.0"
description = "Self-hosted freeform moodboard"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pillow>=11",
    "slowapi>=0.1",
    "python-multipart>=0.0.18",
    "python-dotenv>=1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kemopin"]
